Moved classes out of package objects
- Enabled `-Xlint:package-object-classes`
- Enabled `-Xfatal-warnings`
