--- emma-language/src/test/scala/org/example/foo/package.scala
@@ -1,37 +1,0 @@
-/*
- * Copyright © 2014 TU Berlin (emma@dima.tu-berlin.de)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.example
-
-package object foo {
-
-  def bar: Boolean = true
-
-  class Bar[T](x: Int) {
-    val y: T = null.asInstanceOf[T]
-  }
-
-  object Bar {
-
-    class Baz(val x: Int)
-
-    object Baz {
-      def apply(x: Int): Baz = new Baz(x)
-    }
-
-    def apply(x: Int): Baz = new Baz(x)
-  }
-
-}
