+++ pom.xml
@@ -449,6 +449,8 @@
                         <arg>-Yrangepos</arg>
                         <arg>-Xmax-classfile-name</arg>
                         <arg>140</arg>
+                        <arg>-Xlint:package-object-classes</arg>
+                        <arg>-Xfatal-warnings</arg>
                         <!--<arg>-Xdisable-assertions</arg>-->
                     </args>
                 </configuration>
